backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
boto3==1.34.0
paramiko>=3.4.1
cryptography>=41.0.0
PyYAML==6.0.1
gitpython==3.1.40